move related functions in diags file
