+++ internal/common/conversion/diags.go
@@ -1,6 +1,8 @@
 package conversion
 
 import (
+	"encoding/json"
+
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	sdkv2diag "github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 )
@@ -19,4 +21,31 @@ func FromTPFDiagsToSDKV2Diags(diagsTpf []diag.Diagnostic) sdkv2diag.Diagnostics 
 		})
 	}
 	return results
+}
+
+type ErrBody interface {
+	Body() []byte
+}
+
+// AddJSONBodyErrorToDiagnostics tries to get the JSON body from the error and add it to the diagnostics.
+// For example, admin.GenericOpenAPIError has the Body() []byte method.
+func AddJSONBodyErrorToDiagnostics(msgPrefix string, err error, diags *diag.Diagnostics) {
+	errGeneric, ok := err.(ErrBody)
+	if !ok {
+		diags.AddError(msgPrefix, err.Error())
+		return
+	}
+	var respJSON map[string]any
+	errMarshall := json.Unmarshal(errGeneric.Body(), &respJSON)
+	if errMarshall != nil {
+		diags.AddError(msgPrefix, err.Error())
+		return
+	}
+	errorBytes, errMarshall := json.MarshalIndent(respJSON, "", "  ")
+	if errMarshall != nil {
+		diags.AddError(msgPrefix, err.Error())
+		return
+	}
+	errorJSON := string(errorBytes)
+	diags.AddError(msgPrefix, errorJSON)
 }
